Added SOURCE-ID, STATE, @ and !.
Variables needed from start are now part of Primitive class.
Implementation of a variable becomes easy, though we have not yet
stored any initial value (or even reserved datafield memory for them).

Allocate fixed word identities in a simple and safe way in Dictionary,
and at the same time get the allocation list to be used by words
allocated later.

Implemented QUIT in terms of docol and a mainLoop to be written.