Allow for cross compiling to a polymorphic target.
Use a record with compiler word primitives coupled with a defining
record to build up words being defined.
The target version (cross compiler) compiles to a polymorphic
representation which can be instansiated with different targets.

While this builds, there still remains work on setting up the
target dictionary and figure out how to deal with cell widths.
As before, the target output that is in place will need a lot of
further work as primitives currently are just slices of assembler code
that need proper interface with the Forth machine, which is
non-existent at the moment.