Added IDict and TDict dictionary wrappers.
This was done to ensure we have a single point of 'forall' in FState to
make defining words in the target dictionary polymorphic.

As this increased the complexity in updating deeper record fields and we
must avoid pattern matching in order to keep the polymorphic behavior,
the lens library is introduced.
Using lens library allow for easier updates in nested data structures in
a functional way, also making it easier to write code that keeps the
polymorphic behavior of the target dictionary.

The code compiles but it is still not possible to define new target
words as the compiler record lacks a create function.