Moved unique WordId out from Dictionary and up to FState.
Keep one single unique WordId for every word, no matter whether it is
for the interpreter or target. The word identifiers are used for looking
up the data field on the interpreter and currently not used by targets.
But it is easy to have them there as well and they may find use later.

There was an idea to be able to keep matching WordId's for same word in
the interpreter and target, but that is unlikely to work and I expect it
would be brittle. Better FIND words from both interpreter and target
dictionaries, and later decide which to use. That way the dictionaries
may deviate, and there is really no point to force the interpreter to
have compiled target specific words in order to make them match.