Removed SMUDGE.
Smudge is an old and somewhat unelegant way of making a word available.
It is better to just wait with making the word available until it is done,
and that was what the interpreter did.
Rather than trying to keep this smudge concept, the function is now named
makeAvailable in the interpreter and the intention is to avoid the smudge
concept completely.