Emit token table.
Fixes to get target started with WordId 0.
Keep a mapping from identity number (undressed WordId) to all target
words. This is the basis for the token table.
nameMangle is no longer exported from Translator.Symbol, the idea is to
use addEntityLabel (together with reservedLabels in TargetPrimitive) to
ensure that labels (symbols) are both unique and name mangled.
The "datafields" section (RAM) now uses the "w" flag which seems
appropriate for ELF.