Bind target word labels earlier.
Store the label for target words inside the ForthWord record.
Bind target labels to target words immediately when they are created.
XT now keeps a TargetToken that contains the real target label and the
WordId, which is what the token table index is based on.
Store the Labels record inside the target dictionary record.

This eliminates the previously generated labels for execution tokens
which did not take name mangling and duplicates properly in account.
It also allows the target to easily pick the desired style of index or
label for colon bodies.