Limit names in headers to 31 characters.
It is a good idea to reserve 0 length names for headerless words, in
case we want to provide a means to safely handle them in the same way as
named words. 31 characters should be enough anyway (except perhaps in
cases where we have name mangling, but there are no plans for that at
the moment).