Implemented RECURSE for targets.
Added two functions to TargetPrimitive; branch and branch0.
Implemented recurse when cross compiling.
Added a generic mechanism to inserting local labels in a word. This is
done by adding the instruction offset (instruction count, DOCOL is index
0) to an IntSet. Later when the body is finalized, actual local labels
are inserted.
Added a mechanism to find a named word from the crossCompiler using the
recently factored out dictionary search functionality to Dictionary.