Implemented DO-LOOP for targets.
Compiling loops are now similar to BRANCH and BRANCH0 in that they are
special forms of the compiler and have the branch destination
immediately following.
(LOOP) and (+LOOP) are now part of TargetPrimitive and need to be
implemented with native code. While they are not totally trivial, the
words are part of loops and benefit from being fast. Manipulating the
loop parameters on the rstack is kind of cumbersome in a colon definition.
Fixed +LOOP for interpreter so that if we wrap around (past loop limit),
it is detected and the loop terminates.
Added some library code to target-core.fs
While (DO) is present in target-core.fs, Cortex-M uses native code as it
is short and trivial code.