Implemented DOES> for cross compiler.
Added COMPILER-WORD to specify that the next word is intended to be
compiled in to the interpreter and is meant to be an extension to the
compiler. It is typically meant to be used to define CREATE DOES> words
for the cross compiler.
When encountered, switch to compiling for the interpreter, but keep the
old compiler record available to be restored when the definition is
closed (or we do an ABORT).

When DOES> is reached and we were previously using the cross compiler,
switch to the cross compiler again to compile the DOES> run-time part of
the word for the target. This part is recorded in an nameless word.

The nameless word is compiled into the defining word so that it can
replace the action part of a CREATEd word. By default we put EXIT there
for words that lack DOES>, but it is replaced with the nameless word if
such word is created. To do this, we actually keep two DOES> parts for a
DOHERE, one with EXIT and one that will take the replacing TargetToken
as an argument.

Wrap the ForthWord name into a Maybe to signal that it is a nameless
word.

The code generator will not generate any header for nameless words, nor
will they be part of the word chain. They will however get an execution
token so that they can be used.